ITwitchClient_MessageReceiving extracted
